+++ src/utils/prisma-error-handling.ts
@@ -4,7 +4,7 @@ import {
   PrismaClientUnknownRequestError,
   PrismaClientValidationError,
 } from '@prisma/client/runtime/library';
-import { PrismaException } from 'src/exceptions-filter/prisma-exceptions-filter';
+import { PrismaException } from 'src/exceptions/prisma-exceptions';
 
 export function prismaErrorHandling(error: unknown) {
   if (error instanceof PrismaClientKnownRequestError) {
